requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==43.0.3
Django==5.1.3
djangorestframework==3.15.2
pycparser==2.22
sqlparse==0.5.2
tzdata==2024.2